+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Fixed
+- pyarrow fix did not work on spark executors
+
 ## v0.1.7 (2024-04-12)
 ### Changed
 - Support for pyarrow > 0.17.0